campaign scripts

+++ perf/jmeter/script-it/run_tests.sh
@@ -1,10 +1,54 @@
 #!/usr/bin/env sh
+# set -x
+usage ()
+{
+        test -n "${*}" && echo "${*}"
+    cat <<END_USAGE
+Usage: ${0} {options}
+    where {options} include:
+    -t, --testFile) ** Required
+        relative path to test.json file. 
+    -o, --output
+        relative path of where to publish output.
+    -c, --cooldown
+        time between tests
+    --help
+        Display general usage information
+END_USAGE
+    exit 99
+}
 
-testsJsonFile="tests-template.json"
-test -n "${1}" && testsJsonFile="${1}"
-testsJson=$(cat "$testsJsonFile")
 outputDashboards="dashboard-urls.txt"
-test -n "${2}" && outputDashboards="${2}"
+
+while ! test -z "${1}" ; 
+do
+    case "${1}" in
+        -t|--testFile)
+            shift
+            if test -n "${1}" -a -f "${1}"; then
+              testsJsonFile="${1}"
+            else usage "test file not found"; fi
+            ;;
+        -o|--output)
+            shift
+            if test -n "${1}"; then
+              outputDashboards="${1}"
+            else usage "out file not found"; fi
+            ;;
+        -c|--cooldown)
+            shift
+            if test -n "${1}"; then
+              cooldown="${1}"
+            else usage "out file not found"; fi
+            ;;
+        *)
+            usage "Unrecognized option"
+            ;;
+    esac
+    shift
+done
+
+testsJson=$(cat "$testsJsonFile")
 
 numPd=$(echo "${testsJson}" | jq -r '.pdCount')
 pdIterations=$((numPd -1))
@@ -18,8 +62,12 @@ NAMESPACE=$(echo "${testsJson}" | jq -r ".namespace")
 DURATION=$(echo "${testsJson}" | jq -r ".testDuration")
 export NAMESPACE DURATION
 
+echo "clean leftovers"
+for f in yamls/tmp/* ; do kubectl delete -f $f ; rm $f ; done
+
 for i in $(seq 0 "${testIterations}"); do 
   numThreadGroups=$(echo "${testsJson}" | jq ".tests[$i].threadgroups | length")
+  testId=$(echo "${testsJson}" | jq -r ".tests[$i].id")
   tgIterations=$((numThreadGroups -1))
   for tg in $(seq 0 "${tgIterations}"); do 
     thisTg=$(echo "${testsJson}" | jq -r ".tests[$i].threadgroups[$tg]")
@@ -48,24 +96,46 @@ for i in $(seq 0 "${testIterations}"); do
       testFile="yamls/tmp/test-${i}.yaml"
   done
   
-  echo "clean leftovers"
   kubectl delete -f "${testFile}"
-  startTime=$(date +"%Y-%m-%d %H:%M:%S")
-  startEpoch=$(date -jf "%Y-%m-%d %H:%M:%S" "${startTime}" +%s000)
+  startTime=$(date +"%Y-%m-%dT%H:%M:%S.000Z")
+  startTimeUtc=$(date -u +"%Y-%m-%dT%H:%M:%S.000Z")
+  startEpoch=$(date -jf "%Y-%m-%dT%H:%M:%S.000Z" "${startTime}" +%s000)
   
-  echo "test-$i on: ${testFile} start time ${startTime}"
+  echo "test-$testId on: ${testFile} start time ${startTime}"
   kubectl apply -f "${testFile}"
     echo "letting test run ${testDuration}s"
     sleep "${testDuration}"
 
-  endTime=$(date +"%Y-%m-%d %H:%M:%S")
-  endEpoch=$(date -jf "%Y-%m-%d %H:%M:%S" "${endTime}" +%s000)
-  echo "test-$i on: ${testFile} end time ${endTime}"
+  endTime=$(date +"%Y-%m-%dT%H:%M:%S.000Z")
+  endTimeUtc=$(date -u +"%Y-%m-%dT%H:%M:%S.000Z")
+  endEpoch=$(date -jf "%Y-%m-%dT%H:%M:%S.000Z" "${endTime}" +%s000)
+  echo "test-$testId on: ${testFile} end time ${endTime}"
   test ! -f "${outputDashboards}" && touch "${outputDashboards}"
-  echo "view test-$i results at https://soak-monitoring.ping-devops.com/d/ccanxq9Zk/pingdirectory-performance-test?orgId=1&refresh=5s&from=${startEpoch}&to=${endEpoch}" >> "${outputDashboards}"
-    sleep 3
   
+    sleep 3  
   kubectl delete -f "${testFile}" > /dev/null 2>&1
-  sleep 45
+
+  echo "test-$testId dashboard: https://soak-monitoring.ping-devops.com/d/pdperfrw/pingdirectory-performance-test?orgId=1&from=${startEpoch}&to=${endEpoch}" >> "${outputDashboards}"
+  
+  # returns similar to: /dashboard/snapshot/D2Emm6lWdWM0mnsfSPjcZ8qsj4quN62o
+  echo "snapshotting results"
+  snapshotPath=$(./gen_snapshot.sh "${startTimeUtc}" "${endTimeUtc}" "${testsJsonFile}-${testId}")
+  test -z "${snapshotPath}" && exit 1
+  echo "test-$testId snapshot: https://soak-monitoring.ping-devops.com${snapshotPath}" >> "${outputDashboards}"
+
+  # Store Overall Campaign Results: 
+  test "${i}" -eq 0 && campaignStartTimeUtc="${startTimeUtc}" && campaignStartEpoch="${startEpoch}"
+  if test "${i}" -eq "${testIterations}" ; then
+    campaignEndTimeUtc="${endTimeUtc}"
+    campaignEndEpoch="${endEpoch}"
+    echo "Campaign - $testsJsonFile dashboard: https://soak-monitoring.ping-devops.com/d/pdperfrw/pingdirectory-performance-test?orgId=1&from=${campaignStartEpoch}&to=${campaignEndEpoch}" >> "${outputDashboards}"
+    snapshotPath=$(./gen_snapshot.sh "${campaignStartTimeUtc}" "${campaignEndTimeUtc}" "${testsJsonFile}-${testId}")
+    test -z "${snapshotPath}" && exit 1
+    echo "Campaign - $testsJsonFile snapshot: https://soak-monitoring.ping-devops.com${snapshotPath}" >> "${outputDashboards}"
+  fi
+
+  # cooldown between tests
+  sleep $cooldown
   rm "${testFile}"
+  echo "end of this iteration"
 done
